Made Preconfigurated Version
This WILL break mods that require the configurator, since the function is completely removed. This is an easier form tho, since you do not need a configurator. Once/If it comes out, you will require the Base/UnPreConfigurated version if you wanted to use the CI Configurator (if its ever made)